+++ libraries/mapping/cds_mapper.php
@@ -312,9 +312,9 @@ class Cds_mapper
             $audio_asset->isStreamable = false;
 
             $file_path = $audio_meta->getBaseServerPath() . $audio_meta->file_name;
-            $audio_asset->duration = $audio['audio_duration'] === '' ?
+            $audio_asset->duration = $audio['audio_duration'] === null ?
             $this->get_audio_duration($file_path) :
-            $audio['audio_duration'];
+            intval($audio['audio_duration']);
 
             $audio_enc = new stdClass;
             $audio_enc->href = $audio_guid;